Add files via upload
made the search function more general so that it could be used to search the user table on sign up in case the desired username is already in use. also added an addContacts and searchContacts endpoint as well as updating signup to not make duplicate usernames. all changes have been tested on the server via arc.